Remove all etherpad.com references, mentions, and uses.
The etherpad repository is mostly a direct copy of their site and
infrastructure at the time of the Google acquisition.  There are
a lot of hard-coded etherpad.com references throughout the code.

This changeset removes all etherpad.com specificities in forms of:
  - URLs
  - email addresses (both To and From)
  - corporate specific data (support forum mentions, crazy javascript email
    obfuscator)

Email addresses are now configured in the property file
instead of being hard-coded throughout 40 different files.

A new config value is introduced to specifiy a global To address.
(A config value for a global From address was introduced previously, but
it wasn't used throughout the codebase until now.)

The global To address is used to send automated notices about signups, new
orders, lead gen questions, and exceptional exceptions.