Fixed init.d/etherpad script for debian/ubuntu
For some reason that I cannot yet understand, there are two init.d
scripts with different content. The files are:
  bin/etherpad.init.d
  etherpad/bin/etherpad.init

I rewrote bin/etherpad.init.d to make start, stop and status functions
to work pretty much as intended. The status message currently says
"could not access PID file for etherpad" instead of "etherpad is not
running", when status is queried when etherpad is not running. This
needs to be fixed later.

This change has been tested with Ubuntu 9.10.