Added lintian overrides to silent an invalid error and an invalid warning
The lintian does not correctly detect the magic shell script that is a
scala program in the same time. As a result, lintian will mark this
script as invalid because sh -n reports a parse error. However, this is
not a real parse error, because the script contains unconditional "exit"
before the "error".

The template contains default field for default password which should be
translated. The field contains notes to translators as a comment but
lintian would want to have an extra comment inside the string to translate.
The note as a comment should be enough, make lintian silent about this.