Automatically compute JVM memory limit
The MXRAM (maximum memory for JVM) was hardcoded as 1G (1024 MiB).
Added a piece of bash script that verifies that /usr/bin/perl and
/proc/meminfo are available and then computes a suitable default
value for MXRAM. Default is still "1G" if perl or meminfo are not
available. Any value passed as parameter will still override
this default value (as previously).