Simplify init.d do_stop
The "failsafe" stopping does not work so just drop it. The "--exec $DAEMON"
does not work because we're really looking for a java process, not for a
shell script. However, it's not safe to say "--exec java" either, because
that is identical to "killall java" -- not cool if you have any other java
processes but etherpad.

If "failsafe" stopping is required in the future, it should consist of
grepping output of "jps" or use of pgrep or pkill.