Added code to wait for service to start before returning from init.d script
The init.d script is supposed to return after the service is running if
"start" is passed to the script. Previously the script returned
immediately. This patch makes the script return after the service has been
started or after maximum of 40 seconds (this happens if every try timeouts
in 1 second timeout, normally connection fails immediately and only the
sleep in the init.d script causes delay).