Enable openoffice conversion support by default
Change default setup to enable conversion to MS Word, PDF and OpenDocument.
Note that this patch does not enable import from such formats, because
the file upload functionality is based on cos.jar (http://www.servlets.com/)
which has too weird license.

The etherpad is NOT clever enough to kick the openoffice process going
automatically. It must be started for it beforehand. This patch will start a
headless OpenOffice.org process but it will not monitor it (as far as I know,
OpenOffice.org is not too stable and may crash anytime) or shut it down when
etherpad shuts down.