A status page for the users when the Etherpad instance isn't behaving
This commit adds a status page for when the Etherpad instance should
be automatically restarted by the contrib/bin/etherpad-watcher.sh script.

The status page is an error handler that you just need to setup with your
HTTP gateway, the included configs for apache and nginx has been updated.

The contrib/bin/etherpad-watcher.sh script will write out the last successfull
connection to a javascript-file that the error-handler will include and use to
calculate a restart time.